chore: apply AgentReady improvements (che-23852)
1. Fix tsconfig.json trailing commas so strict-JSON parsers can read it.

2. Add Husky git hooks:
   - .husky/pre-commit — runs lint-staged (eslint + prettier on staged
     files only, replacing the full workspace scan)
   - .husky/commit-msg — runs commitlint to enforce conventional commits

3. Add .commitlintrc.json extending @commitlint/config-conventional
   with lowercase-scope, no-trailing-period, and case-free subject rules.

4. Clean up .gitignore: add editor (*.swo, *.swp, .vscode/), log,
   tsbuildinfo, OS (.DS_Store, Thumbs.db), and coverage/ patterns;
   remove irrelevant binary installer patterns.

5. Add @commitlint/cli, @commitlint/config-conventional, husky, and
   lint-staged to root devDependencies; add lint-staged config.

6. Regenerate .deps license files for new dependencies; fill missing
   MIT/ISC license fields for @commitlint/*, @fastify/*,
   @conventional-changelog/*, and related packages; move unresolved
   pidtree@0.6.1 to EXCLUDED/dev.md; align EXCLUDED/dev.md header
   with EXCLUDED/prod.md; strip prettier column-alignment padding.
