+++ .commitlintrc.json
@@ -1,0 +1,12 @@
+{
+  "extends": ["@commitlint/config-conventional"],
+  "rules": {
+    "type-enum": [2, "always", [
+      "feat", "fix", "docs", "style", "refactor",
+      "perf", "test", "chore", "revert", "ci", "build"
+    ]],
+    "scope-case": [2, "always", "lower-case"],
+    "subject-full-stop": [2, "never", "."],
+    "subject-case": [0]
+  }
+}
