+++ .gitignore
@@ -3,8 +3,6 @@ node_modules/
 .theia/
 .idea/
 
-yarn-error.log
-
 /.yarn/*
 !.yarn/patches
 !.yarn/releases
@@ -19,15 +17,23 @@ run/.che-dashboard-pod
 run/.custom-resources
 run/public-certs
 
-# Exclude binary/installer files
-*.dmg
-*.msi
-*.pkg
-*.deb
-*.rpm
-*.zip
-*.tar.gz
-*.tgz
-
 .claude/settings.local.json
 .worktrees/
+
+# Editor
+*.swo
+*.swp
+.vscode/
+
+# Logs
+*.log
+
+# TypeScript
+*.tsbuildinfo
+
+# OS
+.DS_Store
+Thumbs.db
+
+# Test coverage
+coverage/
