+++ .husky/commit-msg
@@ -1,0 +1,2 @@
+#!/bin/sh
+yarn commitlint --edit "$1"
