+++ packages/dashboard-backend/src/devworkspaceClient/services/__tests__/registryApi.spec.ts
@@ -15,7 +15,6 @@
 import {
   BACKUP_ERROR_CODES,
   BACKUP_IMAGE_DEFAULT_TAG,
-  BackupStatus,
   DEVWORKSPACE_BACKUP_ANNOTATIONS,
   DEVWORKSPACE_BACKUP_LABELS,
 } from '@eclipse-che/common';
@@ -877,164 +876,6 @@ describe('RegistryApiService', () => {
 
       // Workspace found in DW list → workspaceExists: true
       expect(result[0].workspaceExists).toBe(true);
-    });
-  });
-
-  describe('listBackupImages — external registry (quay.io) path', () => {
-    const quayRegistryPath = 'quay.io/my-org/backups';
-
-    interface MockRes {
-      statusCode: number;
-      headers?: Record<string, string>;
-      on: jest.Mock;
-    }
-
-    function buildMockRes(
-      statusCode: number,
-      body: string,
-      headers?: Record<string, string>,
-    ): MockRes {
-      const res: MockRes = {
-        statusCode,
-        headers,
-        on: jest.fn((event: string, handler: (...args: unknown[]) => void) => {
-          if (event === 'data') handler(body);
-          if (event === 'end') handler();
-          return res;
-        }),
-      };
-      return res;
-    }
-
-    /**
-     * Mock the OCI catalog response (direct auth — registry accepts
-     * the provided auth header on _catalog without challenge).
-     * Docker v2 catalog returns full paths including org:
-     * e.g. "my-org/backups/user-che/my-workspace"
-     */
-    function mockCatalogResponse(fullPathRepos: string[]): void {
-      mockHttpsGet.mockImplementationOnce(
-        (_url: unknown, _opts: unknown, callback: (res: MockRes) => void) => {
-          callback(buildMockRes(200, JSON.stringify({ repositories: fullPathRepos })));
-          return { on: jest.fn() };
-        },
-      );
-    }
-
-    beforeEach(() => {
-      mockSaCustomObjectsApi.getNamespacedCustomObject.mockResolvedValue({
-        kind: 'DevWorkspaceOperatorConfig',
-        config: {
-          workspace: {
-            backupCronJob: {
-              registry: { path: quayRegistryPath, authSecret: 'registry-secret' },
-            },
-          },
-        },
-      });
-      const auth = Buffer.from('robot:token').toString('base64');
-      mockCoreV1Api.readNamespacedSecret.mockResolvedValue({
-        data: {
-          '.dockerconfigjson': Buffer.from(
-            JSON.stringify({ auths: { 'quay.io': { auth } } }),
-          ).toString('base64'),
-        },
-      });
-    });
-
-    it('should discover deleted workspace backups from registry', async () => {
-      mockCatalogResponse(['my-org/backups/user-che/deleted-ws']);
-      // No DevWorkspaces exist (workspace was deleted)
-      mockCustomObjectsApi.listNamespacedCustomObject.mockResolvedValue({ items: [] });
-
-      const result = await service.listBackupImages(namespace);
-
-      // Registry confirmed the image exists → always shown even without DW annotation
-      expect(result).toHaveLength(1);
-      expect(result[0].workspaceName).toBe('deleted-ws');
-      expect(result[0].workspaceExists).toBe(false);
-      expect(result[0].timestamp).toBeUndefined();
-      expect(result[0].imageUrl).toBe(
-        `${quayRegistryPath}/${namespace}/deleted-ws:${BACKUP_IMAGE_DEFAULT_TAG}`,
-      );
-    });
-
-    it('should surface deleted workspace with annotation timestamp from registry', async () => {
-      const deletedDWTimestamp = '2026-02-10T12:00:00.000Z';
-      mockCatalogResponse(['my-org/backups/user-che/deleted-ws']);
-      // DW exists with backup annotation but will be marked as not existing
-      // (it's discovered via registry, not DW list)
-      mockCustomObjectsApi.listNamespacedCustomObject.mockResolvedValue({
-        items: [
-          {
-            metadata: {
-              name: 'deleted-ws',
-              annotations: {
-                [DEVWORKSPACE_BACKUP_ANNOTATIONS.LAST_BACKUP_FINISHED_AT]: deletedDWTimestamp,
-              },
-            },
-          },
-        ],
-      });
-
-      const result = await service.listBackupImages(namespace);
-
-      expect(result).toHaveLength(1);
-      expect(result[0].workspaceName).toBe('deleted-ws');
-      expect(result[0].workspaceExists).toBe(true);
-      expect(result[0].timestamp).toBe(deletedDWTimestamp);
-      expect(result[0].imageUrl).toBe(
-        `${quayRegistryPath}/${namespace}/deleted-ws:${BACKUP_IMAGE_DEFAULT_TAG}`,
-      );
-    });
-
-    it('should mark annotation-only entry as UNAVAILABLE when quay confirms image missing', async () => {
-      mockCatalogResponse([]); // quay.io returns empty list
-      mockCustomObjectsApi.listNamespacedCustomObject.mockResolvedValue({
-        items: [
-          {
-            metadata: {
-              name: 'orphaned-ws',
-              annotations: {
-                [DEVWORKSPACE_BACKUP_ANNOTATIONS.LAST_BACKUP_FINISHED_AT]: '2026-02-10T12:00:00Z',
-              },
-            },
-          },
-        ],
-      });
-
-      const result = await service.listBackupImages(namespace);
-
-      expect(result).toHaveLength(1);
-      expect(result[0].workspaceName).toBe('orphaned-ws');
-      expect(result[0].labels[DEVWORKSPACE_BACKUP_ANNOTATIONS.LAST_BACKUP_SUCCESSFUL]).toBe(
-        BackupStatus.UNAVAILABLE,
-      );
-    });
-
-    it('should merge registry and annotation data for existing workspace', async () => {
-      const backupTimestamp = '2026-02-10T14:00:00Z';
-      mockCatalogResponse(['my-org/backups/user-che/my-workspace']);
-      mockCustomObjectsApi.listNamespacedCustomObject.mockResolvedValue({
-        items: [
-          {
-            metadata: {
-              name: 'my-workspace',
-              annotations: {
-                [DEVWORKSPACE_BACKUP_ANNOTATIONS.LAST_BACKUP_FINISHED_AT]: backupTimestamp,
-                [DEVWORKSPACE_BACKUP_ANNOTATIONS.LAST_BACKUP_SUCCESSFUL]: 'true',
-              },
-            },
-          },
-        ],
-      });
-
-      const result = await service.listBackupImages(namespace);
-
-      expect(result).toHaveLength(1);
-      expect(result[0].workspaceName).toBe('my-workspace');
-      expect(result[0].workspaceExists).toBe(true);
-      expect(result[0].timestamp).toBe(backupTimestamp);
     });
   });
 });
